refactor: modularize nerd mode ide layout
